requirements.txt
python-dotenv
langchain 
langchain-community 
langchain-openai 
openai
chromadb 
pymupdf
gradio
sentence-transformers 

ipykernel

python-dotenv
seaborn